Recorder ready.

+++ minisound_ffi/src/include/sound.h
@@ -12,19 +12,8 @@
 typedef struct Sound Sound;
 
 EXPORT Sound *sound_alloc(void);
-
-Result sound_init(
-    Sound *const self,
-    SoundData const sound_data,
-    //
-    // float const *const data,
-    // size_t const data_size,
-    //
-    // SoundFormat const sound_format,
-    // uint32_t const channels,
-    // uint32_t const sample_rate,
-    void *const v_engine
-);
+Result
+sound_init(Sound *const self, SoundData const sound_data, void *const v_engine);
 EXPORT void sound_unload(Sound *const self);
 
 EXPORT Result sound_play(Sound *const self);
@@ -35,7 +24,7 @@ EXPORT void sound_stop(Sound *const self);
 EXPORT float sound_get_volume(Sound const *const self);
 EXPORT void sound_set_volume(Sound *const self, float const value);
 
-EXPORT float sound_get_duration(Sound *const self);
+EXPORT double sound_get_duration(Sound *const self);
 
 EXPORT void
 sound_set_looped(Sound *const self, bool const value, size_t const delay_ms);
